deleteLesson

+++ src/app/Http/Controllers/LessonController.php
@@ -46,10 +46,13 @@ class LessonController extends Controller
 
     /**
      * @param $id
-     * @return mixed
+     * @return \Illuminate\Http\JsonResponse
      */
     public function deleteLesson($id)
     {
-        return $this->lessonService->deleteLessonById($id);
+        if ($this->lessonService->ifLessonExist($id)) {
+            $this->lessonService->deleteLessonById($id);
+        };
+        return response()->json(['result' => 'success'], 201);
     }
 }
